MNT: tax splitting->new module; allow uneven ; cts
Lots of (untested) changes here. Need to test this stuff...
