+++ empress/taxonomy_utils.py
@@ -1,0 +1,129 @@
+import pandas as pd
+
+
+class TaxonomyError(Exception):
+    pass
+
+
+def split_taxonomy(feature_metadata):
+    """ Attempts to find a taxonomy column and split it into taxonomic levels.
+
+    If one of the columns in the feature metadata DataFrame (ignoring case)
+    is named "Taxon" or "Taxonomy", this will return a new DataFrame where
+    the column in question is removed and replaced with S + 1 new columns,
+    where S is the maximum number of semicolons present in any feature's
+    taxonomy.
+
+    Basically, this lets us convert taxonomy annotations to an easier-to-work-
+    with format; instead of having a single column for taxonomy, we'll have one
+    column per taxonomy level. This will make life easier.
+
+    Parameters
+    ----------
+    feature_metadata : pd.DataFrame
+       DataFrame describing feature metadata.
+
+    Returns
+    -------
+    taxsplit_feature_metadata : pd.DataFrame
+        A version of the input feature metadata split as described above. (If
+        none of the columns in the feature metadata were identified as being
+        taxonomy columns, this DataFrame is identical to the input DataFrame.)
+
+    Raises
+    ------
+    TaxonomyError
+        If any of the following conditions are met:
+            1. Multiple possible "taxonomy columns" are present in the input
+               DataFrame: for example, both a "Taxon" and "Taxonomy" column are
+               present, or both a "taxon" and "Taxon" column are present.
+            2. A taxonomy column is present in the input DataFrame, and there
+               is already at least one column in the DataFrame that starts with
+               "Level".
+    """
+    # The entries in this tuple should only be lowercase
+    VALID_TAXONOMY_COLUMN_NAMES = ("taxon", "taxonomy")
+    lowercase_col_names = [str(c).lower() for c in feature_metadata.columns]
+
+    # See if there is a "taxonomy column", and do some related validation on
+    # column names
+    invalid_level_columns_present = False
+    tax_col_index = None
+    tax_col_name = None
+    for col, i in zip(lowercase_col_names, range(len(lowercase_col_names))):
+        if col in VALID_TAXONOMY_COLUMN_NAMES:
+            if tax_col_index is None:
+                tax_col_index = i
+                # ("col" has already been set to lowercase)
+                tax_col_name = feature_metadata.columns[i]
+            else:
+                # Error condition 1 -- multiple possible "taxonomy columns" :(
+                raise TaxonomyError(
+                    (
+                        "Multiple columns in the feature metadata have one of "
+                        "the following names (case insensitive): {}. At most "
+                        "one feature metadata column can have a name from "
+                        "that list."
+                    ).format(VALID_TAXONOMY_COLUMN_NAMES)
+                )
+        if col.startswith("Level"):
+            # This will be a problem *if* there's a taxonomy column, but if
+            # there isn't it's not a problem. (So we wait until after we've
+            # seen all of the feature metadata columns to raise an error about
+            # this.)
+            invalid_level_columns_present = True
+
+    if tax_col_index is not None:
+        if invalid_level_columns_present:
+            # Error condition 2 -- there is at least one "Level" column already
+            # in the feature metadata, which will make distinguishing these
+            # column(s) from the columns we were going to add from feature
+            # metadata difficult.
+            raise TaxonomyError(
+                "The feature metadata contains a taxonomy column, but also "
+                "already contains column(s) starting with the text 'Level'."
+            )
+
+        max_sc_count = 0
+
+        def find_max_semicolon_count(fm_row):
+            # We need to be able to update max_sc_count as we go through each
+            # row, and Python gets angry if we don't explicitly declare it as
+            # "nonlocal" before referencing it here. See
+            # https://stackoverflow.com/a/46018922/10730311
+            nonlocal max_sc_count
+            row_sc_count = fm_row.loc[tax_col_name].count(";")
+            max_sc_count = max(max_sc_count, row_sc_count)
+
+        # Figure out maximum number of semicolons in any feature (this works
+        # on a per-row basis even though axis="columns", because each call
+        # to validate_semicolon_count() gets a Series of columns which is
+        # technically a row I guess)
+        feature_metadata.apply(find_max_semicolon_count, axis="columns")
+
+        # OK, actually do splitting now (taking into account max_sc_count)
+        def split_taxonomy_col(fm_row):
+            levels = [r.strip() for r in fm_row.loc[tax_col_name].split(";")]
+            # If this row's taxonomy has less levels than the max number of
+            # levels, pad it out with empty strings.
+            if len(levels) < max_sc_count + 1:
+                num_missing_levels = (max_sc_count + 1) - len(levels)
+                levels += ["Unspecified"] * num_missing_levels
+            return levels
+
+        # Our use of result_type="expand" means that tax_levels will be a
+        # DataFrame with the same index as feature_metadata but with one column
+        # for each taxonomic level (in order -- Kingdom, Phylum, etc.)
+        tax_levels = feature_metadata.apply(
+            split_taxonomy_col, axis="columns", result_type="expand"
+        )
+        # Assign human-friendly column names: Level 1, Level 2, ...
+        tax_levels.columns = [
+            "Level {}".format(i) for i in range(1, len(tax_levels.columns) + 1)
+        ]
+        fm_no_tax = feature_metadata.drop(columns=tax_col_name)
+        # Finally, join the f.m. with the tax. levels DF by the index.
+        return pd.concat([fm_no_tax, tax_levels], axis="columns")
+    else:
+        # No taxonomy column found, so no need to modify the DataFrame
+        return feature_metadata
