+++ empress/tools.py
@@ -12,10 +12,6 @@ class DataMatchingWarning(Warning):
     pass
 
 
-class FeatureMetadataError(Exception):
-    pass
-
-
 def name_internal_nodes(tree):
     """ Name internal nodes that don't have a name
 
@@ -245,138 +241,3 @@ def match_inputs(
             )
 
     return ff_table, sf_sample_metadata, feature_metadata
-
-
-def split_taxonomy_if_present(feature_metadata):
-    """ Attempts to find a taxonomy column and split it into taxonomic levels.
-
-    If one of the columns in the feature metadata DataFrame (ignoring case)
-    is named "Taxon" or "Taxonomy", this will return a new DataFrame where
-    the column in question is removed and replaced with S + 1 new columns,
-    where S is the number of semicolons present in each feature's taxonomy.
-
-    Basically, this lets us convert taxonomy annotations to an easier-to-work-
-    with format; instead of having a single column for taxonomy, we'll have one
-    column per taxonomy level. This will make life easier.
-
-    Parameters
-    ----------
-    feature_metadata : pd.DataFrame
-       DataFrame describing feature metadata.
-
-    Returns
-    -------
-    taxsplit_feature_metadata : pd.DataFrame
-        A version of the input feature metadata split as described above. (If
-        none of the columns in the feature metadata were identified as being
-        taxonomy columns, this DataFrame is identical to the input DataFrame.)
-
-    Raises
-    ------
-    FeatureMetadataError
-        If any of the following conditions are met:
-            1. Multiple possible "taxonomy columns" are present in the input
-               DataFrame: for example, both a "Taxon" and "Taxonomy" column are
-               present, or both a "taxon" and "Taxon" column are present.
-            2. A taxonomy column is present in the input DataFrame, and there
-               is already at least one column in the DataFrame that starts with
-               "Level".
-            3. A taxonomy column is present in the input DataFrame, but not all
-               features' taxonomies have the same number of semicolons. For
-               example, if one feature has the annotation
-                    "k__;p__;c__;o__;f__;g__;s__"
-               ... and another feature has the annotation
-                    "k__;p__;c__;o__;f__"
-               then this will raise an error, since the taxonomic ranks shared
-               by the two features are not (easily) comparable.
-    """
-    # The entries in this tuple should only be lowercase
-    VALID_TAXONOMY_COLUMN_NAMES = ("taxon", "taxonomy")
-    lowercase_col_names = [str(c).lower() for c in feature_metadata.columns]
-
-    # See if there is a "taxonomy column", and do some related validation on
-    # column names
-    invalid_level_columns_present = False
-    tax_col_index = None
-    tax_col_name = None
-    for col, i in zip(lowercase_col_names, range(len(lowercase_col_names))):
-        if col in VALID_TAXONOMY_COLUMN_NAMES:
-            if tax_col_index is None:
-                tax_col_index = i
-                # ("col" has already been set to lowercase)
-                tax_col_name = feature_metadata.columns[i]
-            else:
-                # Error condition 1 -- multiple possible "taxonomy columns" :(
-                raise FeatureMetadataError(
-                    (
-                        "Multiple columns in the feature metadata have one of "
-                        "the following names (case insensitive): {}. At most "
-                        "one feature metadata column can have a name from "
-                        "that list."
-                    ).format(VALID_TAXONOMY_COLUMN_NAMES)
-                )
-        if col.startswith("Level"):
-            # This will be a problem *if* there's a taxonomy column, but if
-            # there isn't it's not a problem. (So we wait until after we've
-            # seen all of the feature metadata columns to raise an error about
-            # this.)
-            invalid_level_columns_present = True
-
-    if tax_col_index is not None:
-        if invalid_level_columns_present:
-            # Error condition 2 -- there is at least one "Level" column already
-            # in the feature metadata, which will make distinguishing these
-            # column(s) from the columns we were going to add from feature
-            # metadata difficult.
-            raise FeatureMetadataError(
-                "The feature metadata contains a taxonomy column, but also "
-                "already contains column(s) starting with the text 'Level'."
-            )
-
-        # NOTE / TODO: It should be possible to combine validation and
-        # splitting into a single use of apply(). However, not doing this makes
-        # the code simpler (and lets us defer actually creating a new feature
-        # metadata object until we know we need to do so).
-        sc_count = None
-
-        def validate_semicolon_count(fm_row):
-            # We need to be able to update sc_count as we go through each row,
-            # and Python gets angry if we don't explicitly declare sc_count as
-            # "nonlocal" before referencing it here. See
-            # https://stackoverflow.com/a/46018922/10730311
-            nonlocal sc_count
-            row_num_semicolons = fm_row.loc[tax_col_name].count(";")
-            if sc_count is None:
-                sc_count = row_num_semicolons
-            elif sc_count != row_num_semicolons:
-                # Error condition 3 -- inconsistent semicolon counts
-                raise FeatureMetadataError(
-                    "The number of semicolons in the feature metadata's "
-                    "taxonomy column is not consistent for every feature."
-                )
-        # Validate semicolons (this works on a per-row basis even though axis
-        # is "columns", it's because each call to validate_semicolon_count()
-        # gets a Series of columns which is technically a row I guess)
-        feature_metadata.apply(validate_semicolon_count, axis="columns")
-
-        # OK, now we know the taxonomy information is well-formed. We can
-        # actually do splitting now.
-        def split_taxonomy_col(fm_row):
-            return [r.strip() for r in fm_row.loc[tax_col_name].split(";")]
-
-        # Our use of result_type="expand" means that tax_levels will be a
-        # DataFrame with the same index as feature_metadata but with one column
-        # for each taxonomic level (in order -- Kingdom, Phylum, etc.)
-        tax_levels = feature_metadata.apply(
-            split_taxonomy_col, axis="columns", result_type="expand"
-        )
-        # Assign human-friendly column names: Level 1, Level 2, ...
-        tax_levels.columns = [
-            "Level {}".format(i) for i in range(1, len(tax_levels.columns) + 1)
-        ]
-        fm_no_tax = feature_metadata.drop(columns=tax_col_name)
-        # Finally, join the f.m. with the tax. levels DF by the index.
-        return pd.concat([fm_no_tax, tax_levels], axis="columns")
-    else:
-        # No taxonomy column found, so no need to modify the DataFrame
-        return feature_metadata
